feat: new back controller with cors
