+++ ps_accounts.php
@@ -41,6 +41,7 @@ class Ps_accounts extends Module
      */
     private $adminControllers = [
         'AdminAjaxPsAccountsController',
+        'AdminAjaxV2PsAccountsController',
         'AdminDebugPsAccountsController',
         'AdminOAuth2PsAccountsController',
         'AdminLoginPsAccountsController',
